holes in trait parameters
