--- src/Juvix/Compiler/Internal/Translation/FromInternal/Analysis/Traits/Error/Pretty.hs
@@ -1,6 +1,0 @@
-module Juvix.Compiler.Internal.Translation.FromInternal.Analysis.Traits.Error.Pretty
-  ( module Juvix.Compiler.Internal.Translation.FromInternal.Analysis.TypeChecking.Error.Pretty,
-  )
-where
-
-import Juvix.Compiler.Internal.Translation.FromInternal.Analysis.TypeChecking.Error.Pretty
